Change the binary->string converter, add SID, warnings as errors
The existing implementation wasn't really solid and I never liked it
anyway. Egypt rightfully pointed out flaws in the code, so it has
been changed so that it allocates the required memory up front
and returns that string to the caller, who then has to free it up.

Added a function to render SIDS.

Made the build set warnings as errors. Still no idea why this wasn't
the case already.