We need to get the channel hash by reference when modifying the 'data'.
When we read from a channel in PHP, if there is more data returned by
read() than the caller asked for, the data is cached in a 'data' element
in the channel hash. However, since get_channel_by_id() returns a copy,
we immediately lose all of that extra data on the first read. We need to
get the hash by reference in order to modify its elements.