Driver loading in common.lib for sniffer
Implement driver loading in common library via dynamic resolution
of the undocumented NtLoadDriver function and required unicode str
allocator. This method creates a registry entry for the service
which provides userspace interface to driver control, passes that
registry key to the unicode string allocator, in turn passing that
to the driver loading function. Once the driver is started, the
reg key is removed, and at present, the code tries to unlink the
driver file itself. When the driver is to be unloaded, the same
steps are performed except that NtUnloadDriver is called.

Additionally, a resource dropper for the driver files themselves
is provided which reads an embedded VS resource into a buffer and
drops the resulting contents to the local FS. Obviously this isn't
forensically safe - it alters the target in a very loud way. The
sniffer extension has been modified with an embedded resource in
the x86 release version which adds the npf.sys driver into the dll.

This is a WIP commit, does not build at present due to some VS
magic in dealing with changes to common.lib - sniffer builds, but
other extensions fail, and VS believes that terminating semicolons
in C are suddenly a bad thing (source of error is the failing ext
and points to all lines in driver).

Testing:
  None, builds failing