Implement Serializable in java classes
In order for Java to be able to serialize a class, the Serializable
trait (or whatever Java-world calls them) must be implemented in
the class prototype and imported into the source file.

Implement the relevant changes for the common parts of the Java
stagers/payloads in order to permit native Java serialization for
delivery to relevantly deserializing contexts.